fix the bug in training

+++ dataset/utils/collate_fn.py
@@ -63,11 +63,11 @@ def collate_fn(batch):
         # convert to tensor
         frames_1.append(sample[0])
         target_1.append(generate_targets(sample[1], sample[2], sample[3], sample[4], sample[5]))
-        times_1.append(sample[4]*0.1)
+        times_1.append(sample[4].float()*0.1)
 
         frames_2.append(sample[6])
         target_2.append(generate_targets(sample[7], sample[8], sample[9], sample[10], sample[11]))
-        times_2.append(sample[10]*0.1)
+        times_2.append(sample[10].float()*0.1)
 
         similarity_matrix.append(sample[12])
